+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_CodeAgentReminderTool.cs
@@ -28,23 +28,6 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         return Path.Combine([dirName, path]);
     }
     #endregion
-    #region Snippet:Sample_SessionHeaderPolicy_CodeAgentReminderTool
-    private class SessionHeaderPolicy(string agentSessionID) : PipelinePolicy
-    {
-        private static readonly string _SESSION_HEADER = "x-agent-session-id";
-        public override void Process(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-        {
-            message.Request.Headers.Add(_SESSION_HEADER, agentSessionID);
-            ProcessNext(message, pipeline, currentIndex);
-        }
-
-        public override async ValueTask ProcessAsync(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
-        {
-            message.Request.Headers.Add(_SESSION_HEADER, agentSessionID);
-            await ProcessNextAsync(message, pipeline, currentIndex);
-        }
-    }
-    #endregion
     #region Snippet:Sample_CheckRunResult_CodeAgentReminderTool
     protected static void CheckRunResult(RoutineRun completedRun, int minutesWait, bool runCreated)
     {
@@ -79,15 +62,16 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         {
             Versions = { new ProtocolVersionRecord(ProjectsAgentProtocol.Responses, "2.0.0") },
             CodeConfiguration = new(
-                runtime: "python_3_14",
-                entryPoint: ["python", "main.py"],
-                dependencyResolution: CodeDependencyResolution.RemoteBuild
+                runtime: "dotnet_10",
+                entryPoint: ["dotnet", "ToolboxAgent.dll"],
+                dependencyResolution: CodeDependencyResolution.Bundled
             ),
             EnvironmentVariables = {
                 { "AGENT_NAME", middlewareAgentName},
                 { "TOOLBOX_NAME", toolboxName},
                 { "FOUNDRY_PROJECT_ENDPOINT", foundryProjectEndpoint},
                 { "FOUNDRY_MODEL_NAME", modelDeploymentName },
+                { "ASPNETCORE_URLS", "http://+:8088"},
             }
         };
         AgentVersionFromCodeMetadata metadata = new(agentDefinition);
@@ -123,7 +107,7 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         #region Snippet:Sample_CreateAgent_CodeAgentReminderTool_Async
         ProjectsAgentVersion agentVersion = await projectClient.AgentAdministrationClient.CreateAgentVersionFromCodeAsync(
             agentName: "myCodeAgentReminderTool",
-            filePath: GetDirectory(Path.Combine(["Assets", "AgentsCodeToolbox"])),
+            filePath: GetDirectory(Path.Combine(["Assets", "AgentsCodeReminder"])),
             metadata: GetAgentMetadata(
                 middlewareAgentName: "codeAgentMiddleware1",
                 toolboxName: toolBox.Name,
@@ -131,6 +115,8 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
                 modelDeploymentName: modelDeploymentName
             )
         );
+        Console.WriteLine($"Created Agent {agentVersion.Name}, v. {agentVersion.Version}.");
+        Console.WriteLine($"The Agent's identity ID is {agentVersion.InstanceIdentity.ClientId}. Please use it to set \"Foundry User\" permission if needed.");
         #endregion
         #region Snippet:Sample_WaitForDeployment_CodeAgentReminderTool_Async
         while (agentVersion.Status != AgentVersionStatus.Active && agentVersion.Status != AgentVersionStatus.Failed)
@@ -159,10 +145,17 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         #endregion
         #region Snippet:Sample_GetResponseFromAgent_CodeAgentReminderTool_Async
         Console.WriteLine($"Sending prompt {prompt} in session {session.AgentSessionId}...");
-        ProjectOpenAIClientOptions oaiOptions = new();
-        oaiOptions.AddPolicy(new SessionHeaderPolicy(session.AgentSessionId), PipelinePosition.PerCall);
-        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name, options: oaiOptions);
-        ResponseResult response = await responseClient.CreateResponseAsync(prompt);
+        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name);
+        ProjectCreateResponseOptions responseOptions = new()
+        {
+            InputItems = { ResponseItem.CreateUserMessageItem(prompt) },
+            SessionId = session.AgentSessionId,
+        };
+        ResponseResult response = await responseClient.CreateResponseAsync(responseOptions);
+        if (response.Error != null)
+        {
+            throw new InvalidOperationException($"Unable to get the response from an Agent. Error Code: {response.Error.Code}; {response.Error.Message}");
+        }
         Console.WriteLine("Response items:");
         foreach (ResponseItem item in response.OutputItems)
         {
@@ -271,7 +264,7 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         #region Snippet:Sample_CreateAgent_CodeAgentReminderTool_Sync
         ProjectsAgentVersion agentVersion = projectClient.AgentAdministrationClient.CreateAgentVersionFromCode(
             agentName: "myCodeAgentReminderTool",
-            filePath: GetDirectory(Path.Combine(["Assets", "AgentsCodeToolbox"])),
+            filePath: GetDirectory(Path.Combine(["Assets", "AgentsCodeReminder"])),
             metadata: GetAgentMetadata(
                 middlewareAgentName: "codeAgentMiddleware1",
                 toolboxName: toolBox.Name,
@@ -279,6 +272,8 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
                 modelDeploymentName: modelDeploymentName
             )
         );
+        Console.WriteLine($"Created Agent {agentVersion.Name}, v. {agentVersion.Version}.");
+        Console.WriteLine($"The Agent's identity ID is {agentVersion.InstanceIdentity.ClientId}. Please use it to set \"Foundry User\" permission if needed.");
         #endregion
         #region Snippet:Sample_WaitForDeployment_CodeAgentReminderTool_Sync
         while (agentVersion.Status != AgentVersionStatus.Active && agentVersion.Status != AgentVersionStatus.Failed)
@@ -307,10 +302,17 @@ public class Sample_CodeAgentReminderTool : ProjectsOpenAITestBase
         #endregion
         #region Snippet:Sample_GetResponseFromAgent_CodeAgentReminderTool_Sync
         Console.WriteLine($"Sending prompt {prompt} in session {session.AgentSessionId}...");
-        ProjectOpenAIClientOptions oaiOptions = new();
-        oaiOptions.AddPolicy(new SessionHeaderPolicy(session.AgentSessionId), PipelinePosition.PerCall);
-        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name, options: oaiOptions);
-        ResponseResult response = responseClient.CreateResponse(prompt);
+        ProjectResponsesClient responseClient = projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(agentVersion.Name);
+        ProjectCreateResponseOptions responseOptions = new()
+        {
+            InputItems = { ResponseItem.CreateUserMessageItem(prompt) },
+            SessionId = session.AgentSessionId,
+        };
+        ResponseResult response = responseClient.CreateResponse(responseOptions);
+        if (response.Error != null)
+        {
+            throw new InvalidOperationException($"Unable to get the response from an Agent. Error Code: {response.Error.Code}; {response.Error.Message}");
+        }
         Console.WriteLine("Response items:");
         foreach (ResponseItem item in response.OutputItems)
         {